Filter Chart Inputs
Problem: Error when inputting value with preceding "$" causes a problem -
Invalid value The value must be a number (integer or decimal) -
the wording is too advanced for primary school students to understand,
need to implement a filter to prevent such vales from being entered.

Solution: Implement a filter to extract the most probable input the user wants.
If there is not probable input, like "dscsdc", then display error message.

Fixes #4588

Signed-off-by: Agustin Zubiaga <aguz@sugarlabs.org>